nvim: drop unused snippets and cpp_include_picker, make config dir exact_
- snippets を廃止: LuaSnip / friendly-snippets プラグインと自作 snipmate
  スニペット(cpp/go/python)、blink.cmp の snippets ソース・luasnip preset、
  filetype の .snippets マッピングを削除。LSP スニペット展開は blink 内蔵
  エンジンで継続する。
- cpp_include_picker と専用ヘルパ find_dir_in_ancestors、呼び出していた
  <Leader>fh キーマップを削除(get_root は他で使用中のため残す)。
- dot_config/nvim → exact_nvim に rename。~/.config/nvim/ 直下をソースと
  一致するよう pruning し、残骸の snippets/ と空の lsp/ を apply 時に
  自動削除する。
