+++ home/dot_config/exact_nvim/exact_after/exact_ftplugin/c.vim
