+++ home/dot_config/exact_nvim/exact_after/exact_syntax/i3config.vim
