+++ home/dot_config/exact_nvim/exact_lua/exact_armkn/exact_plugins/mini.lua
