--- home/dot_config/nvim/exact_snippets/go.snippets
@@ -1,7 +1,0 @@
-snippet iferr
-  if err != nil {
-    ${0}
-  }
-
-snippet enil
-  err != nil
