--- home/dot_config/nvim/exact_snippets/python.snippets
@@ -1,29 +1,0 @@
-snippet ints
-	[int(s) for s in input().split()]
-
-snippet ii
-	int(input())
-
-snippet s
-	input().strip()
-
-snippet mi
-	map(int, input().split())
-
-snippet ls
-	[input().strip() for _ in range(_)]
-
-snippet ils
-	[int(input()) for _ in range(_)]
-
-snippet graph
-	G: List[List[int]] = [[] for _ in range(N)]
-	for _ in range(M):
-		a, b = map(int, input().split())
-		a -= 1
-		b -= 1
-		G[a].append(b)
-		G[b].append(a)
-
-snippet yes
-	print("Yes" if $1 else "No")
